+++ polygraph/types/tests/test_type_definitions.py
@@ -2,7 +2,8 @@ from unittest import TestCase
 
 from polygraph.types.api import typedef
 from polygraph.types.scalar import Float, Int, String
-from polygraph.types.type_builder import List, NonNull, Union
+from polygraph.types.type_builder import List, NonNull
+from polygraph.types.union import Union
 
 
 class TestTypeDefinition(TestCase):
